Rhbo1767027538828

+++ database/migrations/1766090607702_create_alter_proceding_file_add_is_exclusives_table.ts
@@ -10,8 +10,11 @@ export default class extends BaseSchema {
   }
 
   async down() {
-    this.schema.alterTable(this.tableName, (table) => {
-      table.dropColumn('proceeding_file_type_is_exclusive')
-    })
+    const hasColumn = await this.schema.hasColumn(this.tableName, 'proceeding_file_type_is_exclusive')
+    if (hasColumn) {
+      this.schema.alterTable(this.tableName, (table) => {
+        table.dropColumn('proceeding_file_type_is_exclusive')
+      })
+    }
   }
 }
